Transisiones de niveles

+++ Proyecto/Assets/Scripts/GameController.cs
@@ -2,23 +2,26 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-public class GameController : MonoBehaviour {
+public class GameController : MonoBehaviour
+{
 
     public bool gameOver;
-    public int score;
+    public float score;
     public static GameController instance;
 
     // Use this for initialization
-    void Start () {
+    void Start()
+    {
 
         gameOver = false;
         score = 0;
         instance = this;
 
     }
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
 }
